+++ example/lib/webgl_loader_obj.dart
@@ -181,7 +181,7 @@ class _MyAppState extends State<webgl_loader_obj> {
     if (!kIsWeb) {
       var pars = THREE.WebGLRenderTargetOptions({"format": THREE.RGBAFormat});
       renderTarget =
-          THREE.WebGLRenderTarget((width * dpr), (height * dpr), pars);
+          THREE.WebGLRenderTarget((width * dpr).toInt(), (height * dpr).toInt(), pars);
       renderTarget.samples = 4;
       renderer!.setRenderTarget(renderTarget);
       sourceTexture = renderer!.getRenderTargetGLTexture(renderTarget);
