+++ lib/three3d/renderers/WebGLArrayRenderTarget.dart
@@ -1,7 +1,7 @@
 part of three_renderers;
 
 class WebGLArrayRenderTarget extends WebGLRenderTarget {
-  WebGLArrayRenderTarget(double width, double height, int depth)
+  WebGLArrayRenderTarget(int width, int height, int depth)
       : super(width, height) {
 		this.depth = depth;
     texture = DataArrayTexture(null, width, height, depth);
